Fix "Too many fragment shader texture samplers"
Linux on M1 crash, caused by PBR Terrain. M1 only supports 16 image
samplers in the fragment shader (GL_MAX_TEXTURE_IMAGE_UNITS).
That's why we need to be able to mask RenderTerrainPBRDetail on Linux
too by adding it to an also added TexUnit16orLess list to its feature
table, and to the graphic settings lists above it in the file.