Build Dullahan in Linux aarch64 config stage
GCC needs cstdint header inclusion for it to compile.
include/cef needs to exist first otherwise configuration would fail.
INSTALL_RPATH is needed in try_compile-ing.
PROJECT_ARCH needs to be set on aarch64 to avoid -m64 and
-march(=x86-64) settings which aren't recognised (and wouldn't be
correct) on aarch64.
ENABLE_CXX11_ABI needs to be set ON, otherwise it would use C++ 03's ABI
and cause a linking failure.
Dullahan headers don't seem to be included in the installation upstream,
and dullahan_version.h gets generated only at least after Dullahan
configuration, hence the 2 files copying.
dullahan_host rpath removal is taken out of scope because the Fedora we
support isn't only x86-64 now.
The reindentations are just to make the uniform with the rest in the
file.