Have unencapsulated Expat headers in any condition
The only condition where Expat headers would be encapsulated is when
using LL's Autobuild-based prebuilt libraries, and we're never using
any of LL's prebuilt binary for Expat on desktop, since Expat is
practically available on any supported desktop platform.
The system Expat headers are never encapsulated in any of those
platforms.
This is the beginning of not relying on the LL_USESYSTEMLIBS macro
any more (eventually not relying on the custom USESYSTEMLIBS CMake
setting either).
Keeping the build system still flexible to have the option to use
LL's prebuilt libraries *fully* (we still use some of them in cases
where the platform may not have the related system library or for
convenience, so the term USESYSTEMLIBS may not always be consistent),
is getting harder to maintain.
The way it's done is using #if 1, in order to minimise difference
from upstream.