Not rely on LL_USESYSTEMLIBS for choosing zlib impl
with the effect of choosing zlib over zlib-ng in any condition, which
has been the case so far. But this opens the way to relying on some
other condition, such as which platform, for deciding whether we can
use zlib-ng or not, instead of zlib. So we can freely use, for example
LL's prebuilt, zlib-ng while still using system libraries for others.