Stop trying to force OpenGL version on non-macOS
Turns out my old GPU is detected as complying only to 4.5 when on Linux.
The fact that it's been detected as complying to 4.6 on FreeBSD this
whole time, is quite ironic.
I've only tested that the viewer runs with 4.6 without this on FreeBSD
(the implication works), will test Linux distros next.