Update Windows build instructions
There's a different section for Windows arm64, cause there are
different patterns in the commands that are too difficult to generalise.
BuildTools is used instead of the full Visual Studio IDE,
and BuildTools is installed in Program Files (x86)
on both Windows x64 and arm64.
The CMake used is the one automatically installed by vcpkg, and it
can find the compilers when the toolchain file setting is set to
the file provided by vcpkg.
This way we don't have to install VS CMake component that depends
on x64/x86 build tools, which we don't need for Windows arm64.
The pkg-config used is the other one downloaded, as the path to it
is common on both Windows arm64 and x64.
cURL is installed on arm64 cause there doesn't seem to be any easy
way to build OpenSSL 1.1 for Windows arm64 (which LL's cURL fork
depends on).