Fix SL contributors credits on Windows
Somehow when the stream editing to delete anything that begins with 4
spaces is run from CMake (as opposed to directly on shell), it deletes
most of the contributors on Windows. I've tried many methods like using
\w, \s, etc., but what works is just using the first letters of the bug
codenames. There are still empty results (, , ,) among the final parsed
contributors, but we'll ignore them for now.