LL_WEBRTC for marking WebRTC supporting platforms
so we don't have to keep adding unsupporting ones to the preprocessors
in llvoiceclient.
Note that CM_WEBRTC is complementary to LL_WEBRTC, which means its
purpose is not to be XOR-ed.
Any WebRTC supporting (either using LL's or CM's build) will have
LL_WEBRTC set to ON, but *only* ones that use CM builds will have
CM_WEBRTC set to ON *too*.