Fix crash upon login on Linux aarch64
Tested on Debian Asahi so far, will try on Fedora Asahi next.
I had to trace this down, setting Logging Level to Debug, and comparing
SecondLife.log with a successful login on Linux x86-64.
At first setWeight() in LLViewerVisualParam::setInfo seemed to be
avoiding a crash too when commented out, but turns out it's still okay
when re-enabled.