Connect to Discord now through privacy tab
Now the access token is saved the way passwords are saved, but
without a username, so we can have some persistence without having
to implement an OAuth2 backend server cause we would have to store
those tokens there anyway still and that would even require more
disclosure that the user token gets saved on a server, and it's
just simpler to not go that way. Discord Social SDK doesn't have
to have a helper for sending code to a custom server anyway, that
we would have to have some asynchronous HTTP requestor ready.
Show location check button gets enabled only when Discord
integration is enabled, though it's not functioning yet.