Optimise arrival & departure notifications
by not having extra calls to getAvatars.
The avatars' positions member had to be moved to an object that is
accessible from VOAvatar too, and that would be the global Agent.
It makes sense too, that it's the object that keeps the positions
of other agents. It even has a section for positions too.