Increase chance of arrival notification, correctly
Full (user) name should suffice for the chat log, as it's the one
that is more of a reference anyway instead of display name.
The other avatar's display name is still displayed on the header
anyway. Seems like display names have higher chance of being empty,
that we would miss it being logged/notified because of that.
Also, now that we've optimised the avatars' positions' retrieval,
the numbers can come later after some avatar arrives, that getting
the position using the avatar's ID as the key would result in zero.
In that case, rather than reporting wrong distances (like 403996.2),
it's better to just skip distance information (it shouldn't matter
that much anyway).